+++ src/main/java/opensource/alzheimerdinger/core/domain/notification/entity/FcmToken.java
@@ -1,29 +1,29 @@
-package opensource.alzheimerdinger.core.domain.notification.domain.entity;
+package opensource.alzheimerdinger.core.domain.notification.entity;
 
 import io.hypersistence.utils.hibernate.id.Tsid;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Getter;
 import lombok.NoArgsConstructor;
 import opensource.alzheimerdinger.core.domain.user.domain.entity.User;
 import opensource.alzheimerdinger.core.global.common.BaseEntity;
 
-@Getter
-@Entity
 @Builder
-@AllArgsConstructor
+@Entity
 @NoArgsConstructor
-public class Notification extends BaseEntity {
+@AllArgsConstructor
+public class FcmToken extends BaseEntity {
 
     @Id @Tsid
-    private String notificationId;
-
-    private String title;
-
-    private String body;
+    private String tokenId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
     private User user;
+
+    private String token;
+
+    public void updateToken(String token) {
+        this.token = token;
+    }
 }
