+++ src/main/java/opensource/alzheimerdinger/core/domain/notification/entity/Notification.java
@@ -1,25 +1,29 @@
-package opensource.alzheimerdinger.core.domain.notification.domain.entity;
+package opensource.alzheimerdinger.core.domain.notification.entity;
 
 import io.hypersistence.utils.hibernate.id.Tsid;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 import opensource.alzheimerdinger.core.domain.user.domain.entity.User;
 import opensource.alzheimerdinger.core.global.common.BaseEntity;
 
-@Builder
+@Getter
 @Entity
-@NoArgsConstructor
+@Builder
 @AllArgsConstructor
-public class FcmToken extends BaseEntity {
+@NoArgsConstructor
+public class Notification extends BaseEntity {
 
     @Id @Tsid
-    private String tokenId;
+    private String notificationId;
+
+    private String title;
+
+    private String body;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
     private User user;
-
-    private String token;
 }
