+++ src/main/java/opensource/alzheimerdinger/core/domain/user/domain/service/TokenBlacklistService.java
@@ -27,6 +27,6 @@ public class TokenBlacklistService {
     }
 
     public void blacklist(String token, Duration expiration) {
-        redisTemplate.opsForValue().set(blacklistPrefix + token, "", expiration);
+        redisTemplate.opsForValue().set(blacklistPrefix + token, token, expiration);
     }
 }
