+++ src/main/java/opensource/alzheimerdinger/core/domain/user/ui/UserController.java
@@ -20,6 +20,4 @@ public class UserController {
     public BaseResponse<ProfileResponse> getProfile(@CurrentUser String userId) {
         return BaseResponse.onSuccess(userProfileUseCase.findProfile(userId));
     }
-
-
 }
